NIFI-3413: Incorporated review comments

+++ nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/main/java/org/apache/nifi/processors/standard/db/impl/mysql/event/io/DeleteRowsWriter.java
@@ -74,7 +74,7 @@ public class DeleteRowsWriter extends AbstractBinlogTableEventWriter<DeleteRowsE
             Integer columnType = null;
             if (columnDefinition != null) {
                 jsonGenerator.writeStringField("name", columnDefinition.getName());
-                columnType = (int) columnDefinition.getType();
+                columnType = columnDefinition.getType();
                 jsonGenerator.writeNumberField("column_type", columnType);
             }
             if (row[i] == null) {
