+++ nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/main/java/org/apache/nifi/processors/standard/db/impl/mysql/event/io/InsertRowsWriter.java
@@ -74,7 +74,7 @@ public class InsertRowsWriter extends AbstractBinlogTableEventWriter<InsertRowsE
             Integer columnType = null;
             if (columnDefinition != null) {
                 jsonGenerator.writeStringField("name", columnDefinition.getName());
-                columnType = (int) columnDefinition.getType();
+                columnType = columnDefinition.getType();
                 jsonGenerator.writeNumberField("column_type", columnType);
             }
             if (row[i] == null) {
