+++ nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/main/java/org/apache/nifi/processors/standard/db/impl/mysql/event/io/UpdateRowsWriter.java
@@ -76,7 +76,7 @@ public class UpdateRowsWriter extends AbstractBinlogTableEventWriter<UpdateRowsE
             Integer columnType = null;
             if (columnDefinition != null) {
                 jsonGenerator.writeStringField("name", columnDefinition.getName());
-                columnType = (int) columnDefinition.getType();
+                columnType = columnDefinition.getType();
                 jsonGenerator.writeNumberField("column_type", columnType);
             }
             Serializable[] oldRow = row.getKey();
